cloud_function/requirements.txt
functions-framework>=3.0.0
langchain-core
langchain-text-splitters
langchain-google-vertexai
langchain-google-genai
langchain-mongodb
google-cloud-storage
pymongo[srv]
pypdf
